chat/test: move tests into dedicated test folder
At present we currently have tests in src/, including helpers that are for test purposes only and aren't intended for
use by customers. ably-js also has tests in a dedicated test folder.

This change moves all of our tests to a dedicated test folder.
