+++ test/utils/EventEmitter.test.ts
@@ -1,6 +1,6 @@
 // todo EventEmitter has been copied from @ably/spaces, it's better to move it in separate library
 import { it, describe, expect, beforeEach, afterEach, vi, Mock } from 'vitest';
-import EventEmitter, { removeListener } from './EventEmitter.js';
+import EventEmitter, { removeListener } from '../../src/utils/EventEmitter.js';
 
 describe('removeListener', () => {
   const listener = () => true;
